make components for different sections and import it into page.tsx
