+++ app/_components/GetInTouch.tsx
@@ -1,0 +1,84 @@
+import { useState} from "react"
+import { motion} from "framer-motion"
+import { Button } from "../../components/ui/button"
+import { Input } from "../../components/ui/input"
+import { Textarea } from "../../components/ui/textarea"
+import { toast } from "../../components/ui/use-toast"
+
+
+type GetInTouchProps = {
+    contactRef: React.RefObject<HTMLDivElement>
+}
+
+const GetInTouch = ({ contactRef }: GetInTouchProps) => {
+
+    const [name, setName] = useState("")
+    const [email, setEmail] = useState("")
+    const [message, setMessage] = useState("")
+
+    const handleSubmit = (e: React.FormEvent) => {
+        e.preventDefault()
+        console.log("Form submitted:", { name, email, message })
+        toast({
+            title: "Feedback Submitted",
+            description: "Thank you for your feedback!",
+        })
+        setName("")
+        setEmail("")
+        setMessage("")
+    }
+
+
+    return (
+        <>
+            <motion.section
+                ref={contactRef}
+                className="space-y-6"
+                initial={{ opacity: 0, y: 50 }}
+                whileInView={{ opacity: 1, y: 0 }}
+                transition={{ duration: 0.5 }}
+                viewport={{ once: true }}
+            >
+                <h2 className="text-3xl font-bold">Get in Touch</h2>
+                <form onSubmit={handleSubmit} className="space-y-4">
+                    <div>
+                        <label htmlFor="name" className="block text-sm font-medium text-gray-700">
+                            Name
+                        </label>
+                        <Input
+                            id="name"
+                            value={name}
+                            onChange={(e) => setName(e.target.value)}
+                            required
+                        />
+                    </div>
+                    <div>
+                        <label htmlFor="email" className="block text-sm font-medium text-gray-700">
+                            Email
+                        </label>
+                        <Input
+                            id="email"
+                            type="email"
+                            value={email}
+                            onChange={(e) => setEmail(e.target.value)}
+                            required
+                        />
+                    </div>
+                    <div>
+                        <label htmlFor="message" className="block text-sm font-medium text-gray-700">
+                            Message
+                        </label>
+                        <Textarea
+                            id="message"
+                            value={message}
+                            onChange={(e) => setMessage(e.target.value)}
+                            required
+                        />
+                    </div>
+                    <Button type="submit">Send Feedback</Button>
+                </form>
+            </motion.section>
+        </>
+    )
+}
+export default GetInTouch
