+++ app/layout.tsx
@@ -1,13 +1,13 @@
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
 import '../styles/globals.css'
-
 const inter = Inter({ subsets: ['latin'] })
 
 export const metadata: Metadata = {
   title: 'GDG UPM',
   description: 'Google Developer Group Universiti Putra Malaysia',
 }
+
 
 export default function RootLayout({
   children,
@@ -16,7 +16,13 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <>
+        <body className={inter.className}>
+
+          {children}
+
+        </body>
+      </>
     </html>
   )
 }
